NIFI-11137 remove extra layer from ConsumeJMS
