--- nifi-nar-bundles/nifi-jms-bundle/nifi-jms-processors/src/main/java/org/apache/nifi/jms/processors/strategy/consumer/record/StandardRecordReceivedEventReporter.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.nifi.jms.processors.strategy.consumer.record;
-
-import org.apache.nifi.flowfile.FlowFile;
-import org.apache.nifi.jms.processors.strategy.consumer.EventReporter;
-import org.apache.nifi.processor.ProcessSession;
-
-public class StandardRecordReceivedEventReporter implements EventReporter {
-
-    private final String transitUri;
-
-    public static StandardRecordReceivedEventReporter of(String transitUri) {
-        return new StandardRecordReceivedEventReporter(transitUri);
-    }
-
-    public StandardRecordReceivedEventReporter(String transitUri) {
-        this.transitUri = transitUri;
-    }
-
-    @Override
-    public void reportSuccessEvent(ProcessSession session, FlowFile flowFile) {
-        session.getProvenanceReporter().receive(flowFile, transitUri);
-    }
-}
